skip build if x10c++ already exists

+++ buildx10.sh
@@ -6,10 +6,12 @@ if [ ! -d "x10" ]; then
   unzip x10src.zip
   mv x10-$X10_RELEASE x10
 fi
-cd x10/x10.dist/
-export X10_STATIC_LIB=1
-NPROC=${NPROC:-$(getconf _NPROCESSORS_ONLN)}
-ant -Davailable.procs=$NPROC -Doptimize=true -DNO_CHECKS=true dist
-rm -f lib/libgc.so*
-cd -
+if [ ! -f "x10/x10.dist/bin/x10c++" ]; then
+  cd x10/x10.dist/
+  export X10_STATIC_LIB=1
+  NPROC=${NPROC:-$(getconf _NPROCESSORS_ONLN)}
+  ant -Davailable.procs=$NPROC -Doptimize=true -DNO_CHECKS=true dist
+  rm -f lib/libgc.so*
+  cd -
+fi
